Add changes
- transfer all enum objects into enum.py
- rename mode to override-mode
- transfer merge functions into override.py
- cleanup
