+++ tests/conftest.py
@@ -9,7 +9,7 @@ mount_config = config_fixture('mount-example.yaml')
 endpoint_config = config_fixture('endpoint-example.yaml')
 optional_default_param_config = config_fixture('optional-default-param.yaml')
 pipeline_config = config_fixture('pipeline-example.yaml')
-pipeline_overriden_config = config_fixture('pipeline-overriden-example.yaml')
+pipeline_overridden_config = config_fixture('pipeline-overriden-example.yaml')
 pipeline_with_tasks_config = config_fixture('pipeline-with-tasks-example.yaml')
 pipeline_with_parameters_config = config_fixture('pipeline-with-parameters-example.yaml')
 multiple_param_config = config_fixture('multiple-param.yaml')
