+++ valohai_yaml/objs/pipelines/execution_node.py
@@ -1,33 +1,18 @@
-from enum import Enum
+
 from typing import Any, Dict, List, Optional, Union
 
 from valohai_yaml.lint import LintResult
-from valohai_yaml.objs.pipelines.node import ErrorAction, Node
+from valohai_yaml.objs.enums import ErrorAction, OverrideMode
+from valohai_yaml.objs.pipelines.node import Node
 from valohai_yaml.objs.pipelines.node_action import NodeAction
 from valohai_yaml.types import LintContext, NodeOverrideDict, SerializedDict
-
-
-class Mode(Enum):
-    """What should happen when error occurs in nodes execution."""
-
-    MERGE = 'merge'  # merge inputs in a node with inputs from step
-    OVERRIDE = 'override'  # default: override inputs in a node
-
-    @classmethod
-    def cast(cls, value: Optional[Union['Mode', str]]) -> 'Mode':
-        if not value:
-            return Mode.OVERRIDE
-        if isinstance(value, Mode):
-            return value
-        value = str(value).lower()
-        return Mode(value)
 
 
 class ExecutionNode(Node):
     """Represents an execution node within a pipeline definition."""
 
     type = 'execution'
-    mode: Mode
+    override_mode: OverrideMode
 
     def __init__(
             self,
@@ -37,15 +22,14 @@ class ExecutionNode(Node):
             actions: Optional[List[NodeAction]] = None,
             override: Optional[NodeOverrideDict] = None,
             on_error: Union[str, ErrorAction] = ErrorAction.STOP_ALL,
-
-            mode: Union[str, Mode] = Mode.OVERRIDE
+            override_mode: Union[str, OverrideMode] = OverrideMode.OVERRIDE
     ) -> None:
         super().__init__(name=name, actions=actions, on_error=on_error)
         if override is None:
             override = {}
         self.step = step
         self.override = override
-        self.mode = Mode.cast(mode)
+        self.override_mode = OverrideMode.cast(override_mode)
 
     def lint(self, lint_result: LintResult, context: LintContext) -> None:
         super().lint(lint_result, context)
@@ -65,12 +49,5 @@ class ExecutionNode(Node):
         }
     def get_data(self) -> SerializedDict:
         data = super().get_data()
-        data['mode'] = data['mode'].value
+        data['override_mode'] = data['override_mode'].value
         return data
-
-    def serialize(self) -> SerializedDict:
-        ser = dict(super().serialize())
-        ser['mode'] = self.mode.value
-        if self.mode == Mode.OVERRIDE:
-            ser.pop('mode', None)
-        return ser
